LPD-61412 Move PagesTree to its own file
